Move mocks to separate files
