+++ demoTests/SafetyScreenTestCase.swift
@@ -172,43 +172,3 @@ extension WorldDescription {
                   collisionObjects: collisionObjects)
     }
 }
-
-class MockContainerPresenrer: ContainerPresenter {
-    
-    var currentSafetyState: SafetyState = .none
-    
-    func present(safetyState: SafetyState) {
-        currentSafetyState = safetyState
-    }
-    
-    func presentVision() {}
-    
-    func present(screen: Screen) {}
-    
-    func presentBackButton(isVisible: Bool) {}
-    
-    func present(signs: [ImageAsset]) {}
-    
-    func present(roadDescription: RoadDescription?) {}
-    
-    func present(laneDepartureState: LaneDepartureState) {}
-    
-    func present(calibrationProgress: CalibrationProgress?) {}
-    
-    func dismissMenu() {}
-    
-    func dismissCurrent() {}
-}
-
-class MockAlertPlayer: AlertPlayer {
-    
-    var nowPlayning: AlertSound?
-    
-    func play(sound: AlertSound, repeated: Bool) {
-        nowPlayning = sound
-    }
-    
-    func stop() {
-        nowPlayning = nil
-    }
-}
